feat: Rename circled and solid icons to `IconCircleSolid`
