+++ src/components/Icon/components.tsx
@@ -1,6 +1,6 @@
 export { ReactComponent as Account } from './assets/Account.svg';
 export { ReactComponent as Add } from './assets/Add.svg';
-export { ReactComponent as AddCircle } from './assets/AddCircle.svg';
+export { ReactComponent as AddCircleSolid } from './assets/AddCircleSolid.svg';
 export { ReactComponent as ArrowDown } from './assets/ArrowDown.svg';
 export { ReactComponent as ArrowLeft } from './assets/ArrowLeft.svg';
 export { ReactComponent as ArrowRight } from './assets/ArrowRight.svg';
@@ -21,7 +21,7 @@ export { ReactComponent as ComingSoon } from './assets/ComingSoon.svg';
 export { ReactComponent as Connect } from './assets/Connect.svg';
 export { ReactComponent as Copy } from './assets/Copy.svg';
 export { ReactComponent as Cross } from './assets/Cross.svg';
-export { ReactComponent as Danger } from './assets/Danger.svg';
+export { ReactComponent as CrossCircleSolid } from './assets/CrossCircleSolid.svg';
 export { ReactComponent as Docs } from './assets/Docs.svg';
 export { ReactComponent as Download } from './assets/Download.svg';
 export { ReactComponent as Delete } from './assets/Delete.svg';
@@ -35,7 +35,7 @@ export { ReactComponent as EyeBlocked } from './assets/EyeBlocked.svg';
 export { ReactComponent as Faucet } from './assets/Faucet.svg';
 export { ReactComponent as Globe } from './assets/Globe.svg';
 export { ReactComponent as HamburgerMenu } from './assets/HamburgerMenu.svg';
-export { ReactComponent as InfoCircle } from './assets/InfoCircle.svg';
+export { ReactComponent as InfoCircleSolid } from './assets/InfoCircleSolid.svg';
 export { ReactComponent as Key } from './assets/Key.svg';
 export { ReactComponent as Link } from './assets/Link.svg';
 export { ReactComponent as Lock } from './assets/Lock.svg';
@@ -50,7 +50,7 @@ export { ReactComponent as QRCode } from './assets/QRCode.svg';
 export { ReactComponent as Remix } from './assets/Remix.svg';
 export { ReactComponent as Settings } from './assets/Settings.svg';
 export { ReactComponent as Shield } from './assets/Shield.svg';
-export { ReactComponent as Success } from './assets/Success.svg';
+export { ReactComponent as TickCircleSolid } from './assets/TickCircleSolid.svg';
 export { ReactComponent as Tick } from './assets/Tick.svg';
 export { ReactComponent as TriangleDown } from './assets/TriangleDown.svg';
 export { ReactComponent as TriangleUp } from './assets/TriangleUp.svg';
@@ -58,5 +58,5 @@ export { ReactComponent as Video } from './assets/Video.svg';
 export { ReactComponent as VisaColor } from './assets/VisaColor.svg';
 export { ReactComponent as Wallet } from './assets/Wallet.svg';
 export { ReactComponent as WalletConnect } from './assets/WalletConnect.svg';
-export { ReactComponent as Warning } from './assets/Warning.svg';
+export { ReactComponent as WarningCircleSolid } from './assets/WarningCircleSolid.svg';
 export { ReactComponent as Write } from './assets/Write.svg';
