+++ src/components/Icon/urls.tsx
@@ -1,6 +1,6 @@
 export { default as Account } from './assets/Account.svg';
 export { default as Add } from './assets/Add.svg';
-export { default as AddCircle } from './assets/AddCircle.svg';
+export { default as AddCircleSolid } from './assets/AddCircleSolid.svg';
 export { default as ArrowDown } from './assets/ArrowDown.svg';
 export { default as ArrowLeft } from './assets/ArrowLeft.svg';
 export { default as ArrowRight } from './assets/ArrowRight.svg';
@@ -21,7 +21,7 @@ export { default as ComingSoon } from './assets/ComingSoon.svg';
 export { default as Connect } from './assets/Connect.svg';
 export { default as Copy } from './assets/Copy.svg';
 export { default as Cross } from './assets/Cross.svg';
-export { default as Danger } from './assets/Danger.svg';
+export { default as CrossCircleSolid } from './assets/CrossCircleSolid.svg';
 export { default as Docs } from './assets/Docs.svg';
 export { default as Download } from './assets/Download.svg';
 export { default as Delete } from './assets/Delete.svg';
@@ -35,7 +35,7 @@ export { default as EyeBlocked } from './assets/EyeBlocked.svg';
 export { default as Faucet } from './assets/Faucet.svg';
 export { default as Globe } from './assets/Globe.svg';
 export { default as HamburgerMenu } from './assets/HamburgerMenu.svg';
-export { default as InfoCircle } from './assets/InfoCircle.svg';
+export { default as InfoCircleSolid } from './assets/InfoCircleSolid.svg';
 export { default as Key } from './assets/Key.svg';
 export { default as Link } from './assets/Link.svg';
 export { default as Lock } from './assets/Lock.svg';
@@ -50,7 +50,7 @@ export { default as QRCode } from './assets/QRCode.svg';
 export { default as Remix } from './assets/Remix.svg';
 export { default as Settings } from './assets/Settings.svg';
 export { default as Shield } from './assets/Shield.svg';
-export { default as Success } from './assets/Success.svg';
+export { default as TickCircleSolid } from './assets/TickCircleSolid.svg';
 export { default as Tick } from './assets/Tick.svg';
 export { default as TriangleDown } from './assets/TriangleDown.svg';
 export { default as TriangleUp } from './assets/TriangleUp.svg';
@@ -58,5 +58,5 @@ export { default as Video } from './assets/Video.svg';
 export { default as VisaColor } from './assets/VisaColor.svg';
 export { default as Wallet } from './assets/Wallet.svg';
 export { default as WalletConnect } from './assets/WalletConnect.svg';
-export { default as Warning } from './assets/Warning.svg';
+export { default as WarningCircleSolid } from './assets/WarningCircleSolid.svg';
 export { default as Write } from './assets/Write.svg';
