+++ src/components/ModalState/Icon/styled.tsx
@@ -9,14 +9,14 @@ export const SvgContainer = styled.div`
   display: flex;
 `;
 
-export const Success = styled(Icon.Success)`
+export const Success = styled(Icon.TickCircleSolid)`
   fill: ${({ theme }) => theme.honeycomb.color.success.normal};
 `;
 
-export const Warning = styled(Icon.Warning)`
+export const Warning = styled(Icon.WarningCircleSolid)`
   fill: ${({ theme }) => theme.honeycomb.color.warning.normal};
 `;
 
-export const Danger = styled(Icon.Danger)`
+export const Danger = styled(Icon.CrossCircleSolid)`
   fill: ${({ theme }) => theme.honeycomb.color.danger.normal};
 `;
